Move tests to separate folder structure for clarity
