+++ projects/archiver/test/tasks/indexer/helpers/summary.spec.ts
@@ -1,5 +1,5 @@
 import { IssuePublicationIdentifier, IssueIdentifier } from '../../../../common'
-import { getOtherRecentIssues } from './summary'
+import { getOtherRecentIssues } from '../../../../src/tasks/indexer/helpers/summary'
 
 describe('getOtherRecentIssues', () => {
     const usEdition = 'american-edition'
